backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
xgboost==1.7.6
joblib==1.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7 